requirements.txt
click==8.1.8
decorator==5.1.1
ffmpeg-python==0.2.0
future==1.0.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
moviepy==2.1.1
numpy==2.2.1
opencv-python==4.10.0.84
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
python-dotenv==1.0.1
scenedetect==0.6.5.1
setuptools==75.6.0
tqdm==4.67.1
